Update dependencies and move gradle-versions-plugin configuration to separate file
