+++ gradle-plugin/src/main/kotlin/org/equeim/tremotesf/gradle/GradlePlugin.kt
@@ -4,28 +4,22 @@ import com.android.build.api.dsl.ApplicationExtension
 import com.android.build.api.dsl.CommonExtension
 import com.android.build.api.dsl.LibraryExtension
 import com.android.build.gradle.api.AndroidBasePlugin
-import com.github.benmanes.gradle.versions.updates.DependencyUpdatesTask
 import org.equeim.tremotesf.gradle.utils.*
-import org.equeim.tremotesf.gradle.utils.compileSdk
-import org.equeim.tremotesf.gradle.utils.minSdk
-import org.equeim.tremotesf.gradle.utils.ndk
-import org.equeim.tremotesf.gradle.utils.targetSdk
 import org.gradle.api.JavaVersion
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.artifacts.VersionCatalog
-import org.gradle.api.artifacts.VersionCatalogsExtension
 import org.gradle.api.plugins.ExtensionAware
-import org.gradle.kotlin.dsl.*
+import org.gradle.kotlin.dsl.getByType
+import org.gradle.kotlin.dsl.withType
 import org.jetbrains.kotlin.gradle.dsl.KotlinJvmOptions
 import org.jetbrains.kotlin.gradle.plugin.KotlinAndroidPluginWrapper
-import java.util.*
 
 @Suppress("UnstableApiUsage")
 class GradlePlugin : Plugin<Project> {
     override fun apply(target: Project) {
         target.configureAndroidProject()
-        target.configureVersionsPlugin()
+        target.configureGradleVersionsPlugin()
     }
 
     private fun Project.configureAndroidProject() {
@@ -62,31 +56,5 @@ class GradlePlugin : Plugin<Project> {
 
     private fun KotlinJvmOptions.configureKotlin() {
         jvmTarget = JavaVersion.VERSION_1_8.toString()
-    }
-
-    private fun Project.configureVersionsPlugin() {
-        if (pluginManager.hasPlugin(VERSIONS_PLUGIN_ID)) return
-        pluginManager.apply(VERSIONS_PLUGIN_ID)
-        tasks.named<DependencyUpdatesTask>("dependencyUpdates") {
-            val checker = DependencyVersionChecker()
-            rejectVersionIf {
-                checker.isNonStable(candidate.version)
-            }
-        }
-    }
-
-    private class DependencyVersionChecker {
-        private val stableKeywords = listOf("RELEASE", "FINAL", "GA")
-        private val regex = "^[0-9,.v-]+(-r)?$".toRegex()
-
-        fun isNonStable(version: String): Boolean {
-            val hasStableKeyword = stableKeywords.any { version.toUpperCase(Locale.ROOT).contains(it) }
-            val isStable = hasStableKeyword || regex.matches(version)
-            return isStable.not()
-        }
-    }
-
-    private companion object {
-        const val VERSIONS_PLUGIN_ID = "com.github.ben-manes.versions"
     }
 }
