+++ gradle-plugin/src/main/kotlin/org/equeim/tremotesf/gradle/GradleVersionsPluginConfiguration.kt
@@ -1,0 +1,31 @@
+package org.equeim.tremotesf.gradle
+
+import com.github.benmanes.gradle.versions.updates.DependencyUpdatesTask
+import org.gradle.api.Project
+import org.gradle.kotlin.dsl.named
+import java.util.*
+
+private const val GRADLE_VERSIONS_PLUGIN_ID = "com.github.ben-manes.versions"
+
+fun Project.configureGradleVersionsPlugin() {
+    if (pluginManager.hasPlugin(GRADLE_VERSIONS_PLUGIN_ID)) return
+    pluginManager.apply(GRADLE_VERSIONS_PLUGIN_ID)
+    tasks.named<DependencyUpdatesTask>("dependencyUpdates") {
+        val checker = DependencyVersionChecker()
+        rejectVersionIf {
+            checker.isNonStable(candidate.version)
+        }
+    }
+}
+
+private class DependencyVersionChecker {
+    private val stableKeywords = listOf("RELEASE", "FINAL", "GA")
+    private val regex = "^[0-9,.v-]+(-r)?$".toRegex()
+
+    fun isNonStable(version: String): Boolean {
+        val versionUppercase = version.toUpperCase(Locale.ROOT)
+        val hasStableKeyword = stableKeywords.any(versionUppercase::contains)
+        val isStable = hasStableKeyword || regex.matches(version)
+        return isStable.not()
+    }
+}
