+++ gradle/libs.versions.toml
@@ -1,6 +1,6 @@
 [versions]
 android-gradle-plugin = "7.1.3"
-kotlin = "1.6.20"
+kotlin = "1.6.21"
 gradle-versions-plugin = "0.42.0"
 
 sdk-platform-compile = "31"
@@ -30,7 +30,7 @@ androidx-work = "2.7.1"
 commons-text = "1.9"
 fastscroll = "2.1.0"
 material = "1.5.0"
-publicsuffixlist = "100.0.6"
+publicsuffixlist = "100.0.8"
 timber = "5.0.1"
 threetenabp = "1.4.0"
